+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.4.0] Dec-02-2019
 
 ### Additions
+- [Issue #29] Add send() options parameter
 - [Issue #28] Add npm keywords
 - [Issue #27] Add Mailgun custom options